+++ database/migrations/2025_01_24_163141_add_counter_id_column_in_landlord_agent__auction_bids_table.php
@@ -27,8 +27,10 @@ class AddCounterIdColumnInLandlordAgentAuctionBidsTable extends Migration
      */
     public function down()
     {
-        Schema::table('landlord_agent_auction_bids', function (Blueprint $table) {
-            $table->dropColumn('counter_id');
-        });
+        if (Schema::hasTable('landlord_agent_auction_bids') && Schema::hasColumn('landlord_agent_auction_bids', 'counter_id')) {
+            Schema::table('landlord_agent_auction_bids', function (Blueprint $table) {
+                $table->dropColumn('counter_id');
+            });
+        }
     }
 }
