+++ database/migrations/2025_02_01_150253_add_auction_ended_column_in_landlord_agent_auction_table.php
@@ -27,8 +27,10 @@ class AddAuctionEndedColumnInLandlordAgentAuctionTable extends Migration
      */
     public function down()
     {
-        Schema::table('landlord_agent_auctions', function (Blueprint $table) {
-            $table->dropColumn('auction_ended');
-        });
+        if (Schema::hasTable('landlord_agent_auctions') && Schema::hasColumn('landlord_agent_auctions', 'auction_ended')) {
+            Schema::table('landlord_agent_auctions', function (Blueprint $table) {
+                $table->dropColumn('auction_ended');
+            });
+        }
     }
 }
